Add screen to show Docker disk usage #28
It shows the results of running 'docker system df'. From this screen 'docker system prune'
command can be run (binded to Ctrl+P).

Besides, a new cache library has been added. For now it just being used to hold the results
of running Prune, but it will be used for keeping dry state.