Add initial support for color themes
ColorTheme is now used in most places that render text on the terminal to decide foreground and
background colors. This will allow to easily change between color themes at runtime, but this
is still not supported.

There are currently 3 themes defined: Black256, Dark256 and Light256. The first two are usable,
the light one needs some more polishing.

Currently Black256 is the default color theme.