Add active screen as global variable
Also, added a struct to represent screen dimensions.

Having the screen as a global variable allows to easily access the screen when needed
as compared to having to pass a pointer everywhere is needed. I am not completely
convince it is a good idea, tho, it might just be laziness.