Simplify state management
Container list is now kept by the ContainerStore, which does not care about the position of elements
in the list. This allows to simplify how the container list is managed and the when and how it is
refreshed.

Filter and sorting are now done exclusively at the app layer, the docker layer does not support this
operations.