Change how the decision to render the content of the buffer is taken
Special care has been taken with the new-line callback to avoid using it
after the focus has been lost.

Refreshing signal is controlled with a refresh channel with a buffer of 1, additional
send operations while the channel is full are discarded, since a refresh is going
to happen even if one (or more) refresh signal is lost.