Refactor TextInput
Internally input is kept on a slice array, greatly simplying having to
deal with characters that cannot be represented with a single byte.

Moving around the input now works with well-know keybindings (Ctrl+e to
move to the end, Ctrl+a to move to the start, etc...).

Support for multiline input has been removed, it should not be difficult
to add again, but it it not needed for now.