Check who has the focus on the global refresh call
Without this change, there were situations (like when showing the logs of a container) where
external events that force a refresh were changing the active screen without the user asking for it.

The change seems kind of familiar, which it is probably an indication that how the focus is being
handled is still not fully solved.