Move state to the container widget
State related to what containers have to be shown is now tracked by the
container widget itself. Most (if not all) the logic that decides what
is rendered on the container list screen can now be found on the widget.

Also, the widget now uses the mount/unmount mechanism to update its
content.