Remove container logic and state from Dry struct
The Dry struct will no longer hold all the logic and state of the
application:

* The code that runs a Docker command has a result of a keypress has been
moved to the event handler for the containers screen.

* The state to decide what is shown on the container screen has been moved
to the containers widget.

* Refreshing data now happens when events are received from the
Docker daemon, not periodically.

There is still logic and state for images and networks on it, but it will be
gradually moved as well.