Update Docker dependencies, remove unneeded libraries
Importing the docker/docker and docker/cli projects without setting
constraints in the dep config files had made it very hard to upgrade
versions. This commit should fix the problem, making possible to
both regenerate the vendor directory from scratch and upgrade more
regularly this project dependencies.

This change does not mean that the vendor directory will dissappear
from the repo. It will be recreated and committed in a later change.