Refactor event handling
Now event handlers can decide which handler will handle the next event.
This change simplifies logic to decide who should be handling events by
removing a unique, global, point of coordination that was not working
well on some situations (e.g. the container menu) and it was quite
fragile.

It is also easier now to decide what is going to be shown on the screen.

A thing that still needs to be solved is how to write unit tests for
event handlers.

It would also be good to find a design that will allow to get rid of
global variables.